Convert releases page to typescript
